// src/app/common/models/reservations.ts
export class Reservations {
	_id: string;
	UserId: string;
	FullName: string;
	FirstName: string;
	LastName: string;
	nocheckin: number;
	VisitStatus: string;
	ReservationCode: string;
	RestaurantId: string;
	RestaurantName: string;
	TableId: string;
	TableReservationDate: string;
	TableReservationQuantity: number;
	ModifiedDate: string;
	CreatedDate: string;
}